backend/requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.2
motor>=3.3.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
google-auth>=2.23.0
google-cloud-storage>=2.11.0
python-multipart>=0.0.6
